front_back_interaction/chats/backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Nikita"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
channels = {extras = ["daphne"], version = "^4.1.0"}
channels-redis = "^4.2.0"
celery = "^5.4.0"
djangorestframework = "^3.15.1"
langchain = "^0.1.16"
openai = "^1.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
